Update README to rebuild Jekyll CSS
I believe that the Jekyll rebuild for this repo's previous commit took place while the theme repo (ryanbranch/jekyanbranch) was still updating in some way. As a result, the assets/main.css file being hosted on the web currently is missing an entire new section of "page header nav" styling, despite the rules being present in the SCSS documents for jekyanbranch. Hopefully this commit fixes this odd behavior.