Portfolio Building
Added specific pages for displaying information related to Experience, Software Projects and Hardware Projects
Made a framework for the creation of specific project pages (hardware and software) and "experience pages" (such as a resume). As a test of this, created "project-1.html" and "resume.html" in the new _portfolio_items directory.
Created includes for displaying a type of page nav on a per-page basis dependent on properties related to a new collection called page_types
Created a new collection called "page_types" with a _page_types directory which is used for a sort of "backend tagging system"