Create Gemfile for Local Development (Intermediate Commit)
WARNING: This is an intermediate commit, and some aspects of the site are non-functional. Local builds are currently not working, and I do not necessarily expect this commit to result in a successful build on GitHub.

Ran "bundle init" to create a Gemfile, then added the follow gems within the Gemfile:
 - "jekyll"
 - "jekyll-feed"
 - "jekyll-remote-theme"

Then, updated any dependent gems and brought the site generator to an executable state.

Also created a gitignore which includes the /.jekyll-cache/ directory and the /_site/ folder where generated content is saved