Attempt Remote Rebuild (Intermediate Commit)
WARNING: This is an intermediate commit serving to extend the search for solutions to the issue preventing success in local builds with the jekyanbranch theme.

Added a newline to _config.yml in order to trigger a rebuild. This should remove the '"" -->' from the top of my webpage, but may or may not fix the issues related to local execution.