Undo /gh-pages/ Subfolder Move (Intermediate Commit)
NOTE: The term "Undo" is used in the commit title, but this is NOT a direct revert of the original /gh-pages/ move.

Only after making the file moves in the previous commit (https://github.com/ryanbranch/ryanbranch.github.io/commit/19ab867b00a64f1d145f8eb1931f8ac464c13e81) did I realize that it's not possible to arbitrarily define any directory for Jekyll site generation on GitHub Pages.

As a result, I have made this commit in order to revert to the previous file layout without losing any of the file history information preserved by git.

Also deleted the existing ROOT .gitignore, and updated it (after moving) to match the functionality it held previous to the move, while still retaining the new information.