Create Software Project Pages for Public GitHub Repositories
In this commit I have created a series of new files, all of which directly correspond to a public repository from my GitHub profile.

Page variables have been assigned for each of these 18 new project pages, in order to list the following values in a central table:
 - Project title (WITH LINK to portfolio website page for project)
 - Project creation date
 - Project language(s)
 - Project module(s)
 - Repository branch name (WITH LINK to GitHub page for project)
 - Short description (currently excluded from the table, only displayed on each individual project page)

Also deleted project-1.html, which was a placeholder for the 18 new files added in this commit.