Update Site Copy, Add Google Scholar ID (Intermediate Commit)
WARNING: This is an intermediate commit and some features are not yet complete. Many of the links inserted in about.html do not yet point anywhere. Further, the actual portfolio page copy has not been written.

Updated the site copy on index.html and about.html to be a more valid placeholder for the time being.

Implemented my Google Scholar ID in an attempt to add a profile link within my contact page, corresponding to the recent ryanbranch/jekyanbranch commit.

Also deleted the actions-sandbox.html project page to facilitate its removal, as it is not the type of project which I need to include in my portfolio at this time.

Finally, updated the Gemfile.