Rebuild to fix jekyanbranch Issue #9
Triggering a rebuild of the site in order to test the newest jekyanbranch update, which should list authorship on each blogpost page.

The goal of this commit is to fix the "Blogposts do not list author" Issue (https://github.com/ryanbranch/jekyanbranch/issues/9) in the jekyanbranch repository.

This commit also adds a page.author variable (value "Ryan Branch") for my "Fitting a Plant to Points in Python" blogpost, which may be helpful for testing purposes in the future.