Simplify terms.html
Edited the site copy within terms.html in order to slightly improve simplicity and readability. Clarified certain points, and removed all references of "the Author" in favor of simply mentioning "the Company".

It may still be useful to add some sort of "Summarized Terms" section at the beginning of this page.