Update Site Copy in index.html
This commit serves primarily to regenerate the site, for the specific purpose of confirming viability of the recent Jekyanbranch footer changes.

It also removes the "under construction" phrasing and "More content is coming soon!" message in favor of more detailed information, including links to my GitHub and the two repositories involved in website generation.